refactor sys schema
